resolves userExtractor middleware bug

+++ part4/blog-list/controllers/blogs.js
@@ -1,12 +1,14 @@
 const express = require('express');
 const blogsRouter = express.Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
-const jwt = require('jsonwebtoken');
+const mongoose = require('mongoose');
 
 blogsRouter.get('/', async (_request, response) => {
-  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 });
-  response.json(blogs);
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1, id: 1 });
+
+    response.json(blogs);
 });
 
 blogsRouter.get('/:id', async (request, response, next) => {
@@ -22,13 +24,11 @@ blogsRouter.get('/:id', async (request, response, next) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const { title, author, url, likes} = request.body;
+  const user = request.user;
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-  if (!decodedToken.id) return response.status(401).json({ error: 'token invalid' });
-  const user = await User.findById(decodedToken.id);
-
-  if (!title) return response.status(400).json({error: 'title is missing'});
-  if (!url) return response.status(400).json({error: 'url is missing'});
+  if (!user) return response.status(401).json({error: 'missing token'});
+  if (!title) return response.status(400).json({error: 'missing title'});
+  if (!url) return response.status(400).json({error: 'missing url'});
 
   const blog = new Blog({
     title: title,
@@ -72,7 +72,22 @@ blogsRouter.put('/:id', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:_id', async (request, response, next) => {
+  const user = request.user;
+  if(!user) return response.status(401).json({ error: 'missing token' });
+
+  const validObjectId = mongoose.Types.ObjectId.isValid(request.params._id);
+  if (!validObjectId) {
+    return response.status(400).json({ error: 'Invalid ID format' });
+  }
+
   try {
+    const blog = await Blog.findById(request.params._id);
+    if (!blog) return response.status(404).json({ error: 'blog not found' });
+
+    if (user._id.toString() !== blog.user.toString()) {
+      return response.status(401).json({ error: 'no permission' });
+    }
+
     await Blog.findByIdAndDelete(request.params._id);
     response.status(204).end();
   } catch (error) {
